+++ backend/src/main/java/com/goorm/goormfriends/api/service/UserServiceImpl.java
@@ -25,7 +25,7 @@ public class UserServiceImpl implements UserService{
     private final AuthenticationManagerBuilder authenticationManagerBuilder;
 
     @Override
-    public String findUserEmailByUserId(int userId) throws Exception {
+    public String findUserEmailByUserId(Long userId) throws Exception {
         User user = userRepository.findById(userId).orElseThrow(
                 () -> new Exception("해당 사용자를 찾을 수 없습니다."));
         return user.getEmail();
